requirements.txt
numpy
pandas
plydata
matplotlib
plotnine
mizani
plotly
statsmodels
scikit-learn
xgboost
sklearn-pandas
scikit-misc
fastapi
uvicorn
sqlalchemy
jupyterlab
jupyterlab-server
jupyter-packaging
jupyter-server
ipywidgets
ipympl
jupytext
papermill
nbconvert
streamlit
dash
dash_bootstrap_components
